feat:컬럼 좋아요 기능 구현

+++ src/main/java/jeje/work/aeatbe/repository/AllergyCategoryRepository.java
@@ -3,5 +3,5 @@ package jeje.work.aeatbe.repository;
 import jeje.work.aeatbe.entity.AllergyCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AllergyCategoryRepository extends JpaRepository<AllergyCategory, Integer> {
+public interface AllergyCategoryRepository extends JpaRepository<AllergyCategory, Long> {
 }
