+++ src/main/java/jeje/work/aeatbe/repository/ArticleRepository.java
@@ -3,5 +3,5 @@ package jeje.work.aeatbe.repository;
 import jeje.work.aeatbe.entity.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<Article, Integer> {
+public interface ArticleRepository extends JpaRepository<Article, Long> {
 }
