+++ src/main/java/jeje/work/aeatbe/repository/FreeFromCategoryRepository.java
@@ -3,6 +3,6 @@ package jeje.work.aeatbe.repository;
 import jeje.work.aeatbe.entity.FreeFromCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FreeFromCategoryRepository extends JpaRepository<FreeFromCategory, Integer> {
+public interface FreeFromCategoryRepository extends JpaRepository<FreeFromCategory, Long> {
 
 }
