+++ src/main/java/jeje/work/aeatbe/repository/ProductAllergyRepository.java
@@ -3,6 +3,6 @@ package jeje.work.aeatbe.repository;
 import jeje.work.aeatbe.entity.ProductAllergy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductAllergyRepository extends JpaRepository<ProductAllergy, Integer> {
+public interface ProductAllergyRepository extends JpaRepository<ProductAllergy, Long> {
 
 }
