+++ src/main/java/jeje/work/aeatbe/repository/ProductFreeFromRepository.java
@@ -3,6 +3,6 @@ package jeje.work.aeatbe.repository;
 import jeje.work.aeatbe.entity.ProductFreeFrom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductFreeFromRepository extends JpaRepository<ProductFreeFrom, Integer> {
+public interface ProductFreeFromRepository extends JpaRepository<ProductFreeFrom, Long> {
 
 }
