+++ src/main/java/jeje/work/aeatbe/repository/UserAllergyRepository.java
@@ -3,6 +3,6 @@ package jeje.work.aeatbe.repository;
 import jeje.work.aeatbe.entity.UserAllergy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserAllergyRepository  extends JpaRepository<UserAllergy, Integer> {
+public interface UserAllergyRepository  extends JpaRepository<UserAllergy, Long> {
 
 }
