+++ src/main/java/jeje/work/aeatbe/repository/UserFreeFromRepository.java
@@ -3,6 +3,6 @@ package jeje.work.aeatbe.repository;
 import jeje.work.aeatbe.entity.UserFreeFrom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserFreeFromRepository extends JpaRepository<UserFreeFrom, Integer> {
+public interface UserFreeFromRepository extends JpaRepository<UserFreeFrom, Long> {
 
 }
