+++ src/main/java/jeje/work/aeatbe/repository/WishlistRepository.java
@@ -3,5 +3,5 @@ package jeje.work.aeatbe.repository;
 import jeje.work.aeatbe.entity.Wishlist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WishlistRepository extends JpaRepository<Wishlist, Integer> {
+public interface WishlistRepository extends JpaRepository<Wishlist, Long> {
 }
